requirements.txt
tensorflow>=1.14.0
pillow==6.1.0
SciPy==1.3.1
scikit-image==0.15.0